Fixing Kademlia's K constant.
The current number of 80 defeats the whole purpose of
Kademlia routing, specially since the network is so small,
if we get to have 3 buckets it's probably luck (out of possible 160 buckets
used for routing if the routing has been implemented correctly).

With a K=80, there's very little chance for bucket splits, which
means, each peer is always contacting all the nodes that it knows
for searches. The idea of having buckets which separate peers by
their XOR distance is that we can half the number of searches necessary
if all nodes are organized by these rules.

If we have smaller K sized lists (buckets), we'll start contacting less
peers and getting results a lot faster, as these peers get to ask less peers
in an efficient way.